requirements.txt
cassandra-driver
torch
torchvision